LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-cancel-with-input-expected.txt

CONSOLE MESSAGE: Error: assert_true: input element should be focused expected true got false
Test dialog modal is closed by escape key with input focused



FAIL Test dialog modal is closed by escape key with input focused Error: assert_true: input element should be focused expected true got false
